move librarysearchlist in same component as libraryitemslist
