+++ src/Components/Library/Library.jsx
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import Search from './Search/Search';
 import LibraryItemsList from './LibraryItemsList/LibraryItemsList';
 import CategoriesList from './CategoriesList/CategoriesList';
-import LibrarySearchList from './LibrarySearchList/LibrarySearchList';
 import AddLibraryItem from './Form/AddLibraryItem';
 import LibraryDefault from './LibraryDefault/LibraryDefault';
 import './Library.css';
@@ -24,7 +23,7 @@ const Library = ({ match }) => (
         </div>
         <Switch>
           <Route exact path={`${match.url}/`} component={LibraryDefault} />
-          <Route path={`${match.url}/search`} component={LibrarySearchList} />
+          <Route path={`${match.url}/search`} component={LibraryItemsList} />
           <Route path={`${match.url}/addToLibrary`} component={AddLibraryItem} />
           <Route path={`${match.url}/:category/:type`} component={LibraryItemsList} />
           <Redirect to={`${match.url}/`} />
